pyproject.toml
[tool.poetry]
name = "mcp-woocommerce-operator"
version = "0.1.0"
description = "MCP WooCommerce Operator - AI-powered WooCommerce integration"
authors = ["Idea Bosque"]
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
homepage = "https://github.com/ideabosque/mcp_woocommerce_operator"
repository = "https://github.com/ideabosque/mcp_woocommerce_operator"

[tool.poetry.dependencies]
python = ">=3.8"
httpx = "^0.27.0"
humps = "^0.2.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
respx = "^0.22.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
